+++ lives/src/main/java/com/snakydesign/livedataextensions/Creating.kt
@@ -12,15 +12,15 @@ import android.arch.lifecycle.MutableLiveData
 /**
  * Creates a LiveData that emits the initialValue immediately.
  */
-fun <T> just(initialValue:T): MutableLiveData<T> {
-    return empty<T>().apply { value = initialValue }
+fun <T> liveDataOf(initialValue: T): MutableLiveData<T> {
+    return emptyLiveData<T>().apply { value = initialValue }
 }
 
 
 /**
  * Creates a LiveData that emits the value that the `callable` function produces, immediately.
  */
-fun <T> from(callable : ()->T):LiveData<T>{
+fun <T> liveDataOf(callable: () -> T): LiveData<T> {
     val returnedLiveData = MutableLiveData<T>()
     returnedLiveData.value = callable.invoke()
     return returnedLiveData
@@ -29,6 +29,6 @@ fun <T> from(callable : ()->T):LiveData<T>{
 /**
  * Creates an empty LiveData.
  */
-fun <T> empty():MutableLiveData<T>{
+fun <T> emptyLiveData(): MutableLiveData<T> {
     return MutableLiveData()
 }