+++ lives/src/test/java/com/snakydesign/livedataextensions/Utils.kt
@@ -1,0 +1,7 @@
+package com.snakydesign.livedataextensions
+
+open class TestOnNextAction<T> : OnNextAction<T> {
+    override fun invoke(p1: T?) {
+    }
+
+}